refactor(core): split carrier/composition typeclasses into accessor, forgetful, compose packages
The 11 typeclasses that lived directly in `dev.constructive.eo` (and the
accessor trio that was awkwardly in `eo.data`) now sit in three packages by
responsibility:

- eo.accessor   — Accessor, PartialAccessor, ReverseAccessor
                  (focus extract / inject primitives)
- eo.forgetful  — ForgetfulFunctor, ForgetfulApplicative, ForgetfulFold,
                  ForgetfulTraverse (cats-lifted focus operations)
- eo.compose    — AssociativeFunctor, Composer, Morph, ReadCompose
                  (optic-combination machinery; all import Optic)

The accessor/forgetful boundary is the read-primitive vs cats-operation split;
the compose boundary is objective — those four are exactly the typeclasses that
depend on Optic. Choice B of the three layouts considered.

These were visible to optics/ and data/ for free via enclosing-package scope
and to downstream modules via `import dev.constructive.eo.X`; the move makes
every use site import explicitly. No behaviour change — pure namespacing.
Full suite green, CompositionMatrixSpec still 100/100, benchmarks compile.
Doc snippets (optics.md, multifocus.md) repointed too.
