+++ core/src/main/scala/dev/constructive/eo/accessor/Accessor.scala
@@ -1,5 +1,5 @@
 package dev.constructive.eo
-package data
+package accessor
 
 /** Read the focus `A` out of a two-parameter carrier `F[X, A]`. Required by `Optic.get`;
   * always-hitting carriers (`Tuple2` for Lens, `Direct` for Iso / Getter) supply one.
