+++ core/src/main/scala/dev/constructive/eo/compose/AssociativeFunctor.scala
@@ -1,4 +1,5 @@
 package dev.constructive.eo
+package compose
 
 import cats.syntax.either.*
