+++ core/src/main/scala/dev/constructive/eo/compose/Composer.scala
@@ -1,4 +1,5 @@
 package dev.constructive.eo
+package compose
 
 import optics.Optic
