+++ core/src/main/scala/dev/constructive/eo/compose/Morph.scala
@@ -1,4 +1,5 @@
 package dev.constructive.eo
+package compose
 
 import optics.Optic
 import data.Affine
