+++ core/src/main/scala/dev/constructive/eo/forgetful/ForgetfulFold.scala
@@ -1,4 +1,5 @@
 package dev.constructive.eo
+package forgetful
 
 import cats.Monoid
