+++ laws/src/main/scala/dev/constructive/eo/laws/data/AffineLaws.scala
@@ -2,7 +2,7 @@ package dev.constructive.eo.laws.data
 
 import cats.{Applicative, Id}
 import dev.constructive.eo.data.Affine
-import dev.constructive.eo.{ForgetfulFunctor, ForgetfulTraverse}
+import dev.constructive.eo.forgetful.{ForgetfulFunctor, ForgetfulTraverse}
 
 /** Carrier-level laws for `Affine[X, A]`.
   *
