+++ laws/src/main/scala/dev/constructive/eo/laws/data/discipline/SetterFTests.scala
@@ -1,7 +1,7 @@
 package dev.constructive.eo.laws.data.discipline
 
-import dev.constructive.eo.ForgetfulFunctor
 import dev.constructive.eo.data.{Fst, SetterF, Snd}
+import dev.constructive.eo.forgetful.ForgetfulFunctor
 import dev.constructive.eo.laws.data.SetterFLaws
 import org.scalacheck.Prop.forAll
 import org.scalacheck.{Arbitrary, Cogen}
