+++ laws/src/main/scala/dev/constructive/eo/laws/eo/FoldAndTraverseLaws.scala
@@ -2,6 +2,8 @@ package dev.constructive.eo
 package laws
 package eo
 
+import dev.constructive.eo.forgetful.*
+
 import optics.Optic
 import optics.Optic.*
